Rework Textures
Textures have been changed to be more reuseable, they now live in thier
own module (Textuers.elm) rather than in the View and are comprised of a
base path ("..\\textuers\\") and a specific path related to the Kind of
Tile (e.g. "dirt.jpg"). Highlighting and selecting have been changed to
add "h-" and "s-" respectively to these urls to select the correct
texture. Examples of this can be found in Main and View. I have also
added a Queen Tile to the model wich currently reperesent the location
of the queen only and displays a blank square.